pyproject.toml
[project]
name = "online-store"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.16.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]==0.115.14",
    "pydantic-settings==2.10.1",
    "sqlalchemy[asyncio]==2.0.41",
]

[dependency-groups]
dev = [
    "mypy==1.16.1",
    "ruff==0.12.2",
]
